dom/base/test/test_bug393968.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=393968
-->
<head>
  <title>Test for Bug 393968</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=393968">Mozilla Bug 393968</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug 393968 **/
var req = new XMLHttpRequest();
req.open("POST", window.location.href);
req.setRequestHeader("Content-Type", "text/plain; charset=us-ascii; boundary=01234567890");
req.send("Some text");

is(SpecialPowers.wrap(req).channel
      .QueryInterface(SpecialPowers.Ci.nsIHttpChannel)
      .getRequestHeader("Content-Type"),
   "text/plain;charset=UTF-8;boundary=01234567890",
   "Headers should match");
</script>
</pre>
</body>
</html>
